Update read_key_file for variable-sized keys

+++ src/streamfile.h
@@ -155,7 +155,7 @@ STREAMFILE * open_stream_name(STREAMFILE *streamFile, const char * ext);
 
 int read_string(char * buf, size_t bufsize, off_t offset, STREAMFILE *streamFile);
 
-int read_key_file(uint8_t * buf, size_t bufsize, STREAMFILE *streamFile);
+size_t read_key_file(uint8_t * buf, size_t bufsize, STREAMFILE *streamFile);
 int read_pos_file(uint8_t * buf, size_t bufsize, STREAMFILE *streamFile);
 
 int check_extensions(STREAMFILE *streamFile, const char * cmp_exts);
